requirements.txt
Django==1.11
django-cors-headers==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
mysqlclient==1.4.6
pytz==2019.3
